Add get_parameter_query function in uri class
A new function, get_parameter_query(), has been added to the uri class to consolidate and simplify URI parameter operations. This function generates a string of URI parameters instead of replicating this logic in multiple places, which promotes code reuse and maintains cleaner code. The existing codes that perform the same task have been replaced with the call to this new function.