Update CMakeLists.txt and change the project structure
The project structure has been changed with the creation of a new CLI executable (cnet-cli). This refactoring also includes replacing the 'example' project with 'cnet-cli'. Furthermore, the libraries are now linked publicly and the build type is checked for a release configuration to apply certain compile options. Lastly, OpenSSL libraries are now found and linked to 'cnet' project.