Update CMakeLists.txt for cnet-cli
This commit simplifies CMakeLists.txt for the `cnet-cli` target. It removes the condition for Release build type which managed specific compile options and also removed the OpenSSL package inclusion and linking. Furthermore, it specifies that cnet-cli target should include directories under 'vendor' and compiles with `-Os` option for better optimization.