Remove OpenSSL dependency from tcp_client
The OpenSSL dependency within the tcp_client has been removed. This change involves the removal of the include to "openssl/ssl.h.in", and the removal of SSL related operations in the destructor and tcp_client function. This will streamline the tcp_client operations, and reduce overhead in not needing OpenSSL.